adds restriction that only logged in users can delete reviews

+++ routes/reviews.js
@@ -5,9 +5,16 @@ router.get('/:reviewID/delete', (request, response) => {
   const reviewID = request.params.reviewID
   const userID = request.session.userID
 
-  database.deleteReview(reviewID, error => {
-    response.redirect('/users/' + userID)
-  })
+  if (!userID) {
+    response.render('error', { error: {
+      message: 'You must be logged in to delete a review' },
+      userID: request.session.userID
+    })
+  } else {
+    database.deleteReview(reviewID, error => {
+      response.redirect('/users/' + userID)
+    })
+  }
 })
 
 module.exports = router
